feat: Text Editor 생성 refactor: Dropdown
