+++ package.json
@@ -11,9 +11,11 @@
   },
   "dependencies": {
     "axios": "^1.8.2",
+    "quill": "^2.0.3",
     "react": "^19.0.0",
     "react-dom": "^19.0.0",
     "react-icons": "^5.5.0",
+    "react-quill": "^2.0.0",
     "react-router-dom": "^7.2.0",
     "react-toastify": "^11.0.5",
     "rolling-paper": "file:",
